+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 - Include command name in unsupported container command error messages (#1653)
+- Updated arguments name for GCRA command (#1667)
 
 ### Fixed
 - Fixed handling of gap slots in `SlotMap::offsetUnset()` (#1660)